Refactor physics: separate geometry, unify single/multi electrode APIs
- Extract physics tests into tests/unit/test_physics.py
- get_coeffs_line_source and get_coeffs_point_source now accept (3,) or
  (N_elec, 3) electrode positions — no separate batch functions needed
- SegmentGeometry is parsed once by the caller and passed in
- Memory chunking moved to get_weights (electrode_chunk_size parameter)
- Add log_rank0 helper to utils.py, use consistently
- Expose --electrode-chunk-size in CLI
- Add multi-electrode + varying sigma test for point source
