+++ tests/integration/test_weights.py
@@ -170,7 +170,7 @@ def test_rat_s1_interleaved_electrode_types(tmp_path):
     per-electrode sigma values (0.3 and 0.4). Compares the batched result
     against computing each electrode individually.
     """
-    from bluerecording.physics import get_coeffs_line_source, get_coeffs_point_source
+    from bluerecording.physics import SegmentGeometry, get_coeffs_line_source, get_coeffs_point_source
     from tests.conftest import EXAMPLE_RAT_S1
 
     circuit_config = str(EXAMPLE_RAT_S1 / "circuit_config.json")
@@ -193,11 +193,12 @@ def test_rat_s1_interleaved_electrode_types(tmp_path):
     from bluerecording.weights import _get_segment_midpts
 
     mid_positions = _get_segment_midpts(pos_df, node_ids)
+    geom = SegmentGeometry.from_positions(pos_df)
 
     for i, electrode in enumerate(electrodes):
         s = sigma[i]
         if electrode.type.value == "LineSource":
-            expected = get_coeffs_line_source(pos_df, columns, electrode.position, s)
+            expected = get_coeffs_line_source(geom, columns, electrode.position, s)
         else:
             expected = get_coeffs_point_source(mid_positions, electrode.position, s)
